This is a port of existing code, excluding anything related to quizzes and user login; the corresponding code is commented out. Quiz-related props still exist, but they're unused, and dummy values are used in some places.
The directory public is empty (except for collection.md) and will stay this way.
For the time being, symlink the existing `books` directory into `public`.

The code uses the latest versions of dependencies, except for tailwind, for which I
had to downgrade to version 3 after spending a few hours trying to mediate the fight
between tailwind and sass. As a result of this fight, CSS's are reorganized a bit,
which may be unnecessary.

React code is changed in a few places. The two I remember (and which may be the
only two overal) are:
- React.ReactNode must be used in instead of JSX.Element
- useRef now requires initial value and type (if initial value is null).

This code survives `yarn build` and the rendering looks correct.